requirements.txt
Django~=5.0
djangorestframework~=3.16
gunicorn>=21.2
psycopg2-binary>=2.9
django-environ>=0.11.2
whitenoise>=6.6
dj_database_url
drf-spectacular[sidecar]
celery
simpleeval
redis
django-cors-headers

pytest
pytest-django
pytest-cov
requests

black==25.9.0
isort==7.0.0
